# packages/api-client/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/milk-records": {
      "post": {
        "operationId": "MilkRecordsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMilkRecordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilkRecordDto"
                }
              }
            }
          }
        },
        "tags": [
          "milk-records"
        ]
      },
      "get": {
        "operationId": "MilkRecordsController_findAll",
        "parameters": [
          {
            "name": "animalId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MilkRecordDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "milk-records"
        ]
      }
    },
    "/milk-records/{id}": {
      "get": {
        "operationId": "MilkRecordsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilkRecordDto"
                }
              }
            }
          }
        },
        "tags": [
          "milk-records"
        ]
      },
      "patch": {
        "operationId": "MilkRecordsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilkRecordDto"
                }
              }
            }
          }
        },
        "tags": [
          "milk-records"
        ]
      },
      "delete": {
        "operationId": "MilkRecordsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilkRecordDto"
                }
              }
            }
          }
        },
        "tags": [
          "milk-records"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dairy Farm API",
    "description": "Dairy Farm Management API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateMilkRecordDto": {
        "type": "object",
        "properties": {
          "animalId": {
            "type": "string",
            "example": "animal_123"
          },
          "farmId": {
            "type": "string",
            "example": "farm_123"
          },
          "date": {
            "type": "string",
            "example": "2026-03-10"
          },
          "session": {
            "type": "string",
            "enum": [
              "MORNING",
              "EVENING"
            ]
          },
          "quantity": {
            "type": "number",
            "example": 8.5
          },
          "fatContent": {
            "type": "number",
            "example": 4.2
          },
          "quality": {
            "type": "string",
            "enum": [
              "EXCELLENT",
              "GOOD",
              "AVERAGE",
              "POOR"
            ]
          },
          "notes": {
            "type": "string",
            "example": "Good yield today"
          }
        },
        "required": [
          "animalId",
          "farmId",
          "date",
          "session",
          "quantity"
        ]
      },
      "MilkRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "farmId": {
            "type": "string"
          },
          "animalId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "session": {
            "type": "string",
            "enum": [
              "MORNING",
              "EVENING"
            ]
          },
          "quantity": {
            "type": "number"
          },
          "fatContent": {
            "type": "number"
          },
          "quality": {
            "type": "string",
            "enum": [
              "EXCELLENT",
              "GOOD",
              "AVERAGE",
              "POOR"
            ]
          },
          "notes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "farmId",
          "animalId",
          "date",
          "session",
          "quantity",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}